refactor: introduce is_external_failed to PipelineData, and simplify try logic
